apps/facial_recognition_system/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pyjwt>=2.8.0

# Database
sqlalchemy>=2.0
pymysql>=1.1.0
cryptography>=41.0.0

# Environment and configuration
python-dotenv>=1.0.0

# Templating
jinja2>=3.1.0

# Face recognition and computer vision
insightface>=0.7.3
onnxruntime>=1.16.0
opencv-python>=4.8.0
mediapipe>=0.10.0

# Numerical and image processing
numpy>=1.24.0
Pillow>=10.0.0
scipy>=1.11.0
